move some types into type file to allow mac and linux builds
